app/requirements.txt
asgiref==3.5.2
Django==3.2.6
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
gunicorn==20.1.0
Pillow==2.2.1
PyJWT==2.6.0
pytz==2022.4
psycopg2-binary==2.9.1
sqlparse==0.4.3
